closes #383 - put splitWithSearchTerm as a helper function, made test
